RAGAgent/requirements.txt

langchain
langgraph
langchainhub
langchain-community
sentence_transformers
langchain_pinecone
langchain_groq
duckduckgo-search
python-dotenv
black
isort

# langchain_openai
# tavily-python
# beautifulsoup4
